Move block body to its own module
