+++ chainstate/test-framework/src/block_builder.rs
@@ -20,8 +20,9 @@ use crate::utils::{create_multiple_utxo_data, create_new_outputs, outputs_from_b
 use crate::TestFramework;
 use chainstate::{BlockSource, ChainstateError};
 use chainstate_types::BlockIndex;
+use common::chain::block::block_body::BlockBody;
 use common::chain::block::signed_block_header::{BlockHeaderSignature, BlockHeaderSignatureData};
-use common::chain::block::{BlockBody, BlockHeader};
+use common::chain::block::BlockHeader;
 use common::chain::{OutPoint, OutPointSourceId};
 use common::{
     chain::{
